requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
bleach==3.3.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-constance==4.1.3
django-summernote==0.8.20.0
django-timezone-field==7.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
GDAL==3.3.0
idna==3.10
kombu==5.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
XlsxWriter==3.2.0
